+++ stack--2022/3-advanced--multi/soft-execution-context/src/root.ts
@@ -1,5 +1,3 @@
-import { getGlobalThis } from '@offirmo/globalthis-ponyfill'
-
 import { SoftExecutionContext } from './types.js'
 import { LIB } from './consts.js'
 import { _createSEC } from './core.js'
@@ -7,7 +5,7 @@ import { _createSEC } from './core.js'
 /////////////////////
 
 function getRootSEC<Injections = {}, AnalyticsDetails = {}, ErrorDetails = {}>(): SoftExecutionContext<Injections, AnalyticsDetails, ErrorDetails> {
-	const global_this = getGlobalThis<any>()
+	const global_this = globalThis as any
 
 	if (!global_this.__global_root_sec) {
 		//console.log(`[${LIB}] Creating root context…`)
@@ -18,7 +16,7 @@ function getRootSEC<Injections = {}, AnalyticsDetails = {}, ErrorDetails = {}>()
 }
 
 function _test_only__reset_root_SEC() {
-	const global_this = getGlobalThis<any>()
+	const global_this = globalThis as any
 
 	delete global_this.__global_root_sec
 }