+++ stack--2022/3-advanced--multi/universal-debug-api--placeholder/src/index.ts
@@ -1,18 +1,16 @@
-import { getGlobalThis } from '@offirmo/globalthis-ponyfill'
 import { DebugApiRoot, DebugApi } from '@offirmo/universal-debug-api-interface'
 
 import createV1 from './v1.js'
 
 // ensure the root is present
 const _p = '_debug'
-const globalThis = getGlobalThis<any>()
-globalThis[_p] ||= {} as DebugApiRoot
+;(globalThis as any)[_p] ||= {} as DebugApiRoot
 
 // install globally if no other implementation already present
-globalThis[_p].v1 ||= createV1()
+;(globalThis as any)[_p].v1 ||= createV1()
 
 // expose the installed implementation
-const instance: DebugApi = globalThis[_p].v1
+const instance: DebugApi = (globalThis as any)[_p].v1
 
 const {
 	getLogger,
@@ -26,8 +24,6 @@ export {
 	exposeInternal,
 	overrideHook,
 	addDebugCommand,
-
-	globalThis, // for convenience
 
 	createV1, // for special cases
 }