+++ stack--2022/3-advanced--multi/universal-debug-api--placeholder/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [feat] fixed compatibility with the latest Typescript version (4.9)
 * [chore][breaking] converted to full ESM, exposed as `"type": "module"`, requires node 12+, node 18+ recommended
 * [doc] marked as maintained in 2024! Happy new year!
+* [chore] removed usage of a globalThis ponyfill, this feature is now widely supported
 * ...
 
 ## v1.0.1